+++ projects/net.splitcells.gel.core/src/main/java/net/splitcells/gel/problem/ProblemI.java
@@ -275,4 +275,9 @@ public class ProblemI implements Problem {
     public Node toDom() {
         return this.allocations.toDom();
     }
+
+    @Override
+    public Object identity() {
+        return this;
+    }
 }