+++ projects/net.splitcells.gel.sheath/src/main/java/net/splitcells/gel/solution/SolutionAspect.java
@@ -290,4 +290,9 @@ public class SolutionAspect implements Solution {
     public String toString() {
         return path().toString();
     }
+
+    @Override
+    public Object identity() {
+        return solution.identity();
+    }
 }